fix : CarFactory 클래스 삭제
